test_timezone() split to utils test file
